+++ CHANGELOG.md
@@ -7,7 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-* None.
+### Added
+
+* Add `LogLevel.DEBUG` and debug log in `Parser`.
+* Add logLevel parameter for `Parser.constructor`.
+* Add error code for `Cryptor` internal error.
+
+### Changed
+
+* Change `Logger.logLevel` default. (`error` => `warning`)
+* Rename `LogLevel.WARNING` to `LogLevel.WARN`.
 
 ## [0.4.1 (2019-02-12)]